+++ src/store/config.ts
@@ -1,6 +1,5 @@
+import { BaseConfig } from "../common/base-config";
 import { Tree } from "@angular-devkit/schematics/src/tree/interface";
-import { Workspace } from "../interfaces/workspace.interface";
-import { Project } from "../interfaces/project.interface";
 
 export interface Schema {
     name: string;
@@ -11,23 +10,7 @@ export interface TemplateConfig extends Omit<Schema, "path"> {
     prefix?: string;
 }
 
-export class Config implements Schema {
-    private workspace: Workspace | null;
-    private project: Project | null;
-
-    private get projectSourcePath(): string | null {
-        if (!this.project) return null;
-
-        const root = this.project.sourceRoot
-            ? `/${this.project.sourceRoot}`
-            : `/${this.project.root}/src/`;
-        const projectDirName =
-            this.project.extensions["projectType"] === "application"
-                ? "app"
-                : "lib";
-        return `${root}${projectDirName}`;
-    }
-
+export class Config extends BaseConfig implements Schema {
     public get name(): string {
         return this.schema.name;
     }
@@ -43,26 +26,7 @@ export class Config implements Schema {
         };
     }
 
-    constructor(private readonly schema: Schema, private readonly tree: Tree) {
-        this.load();
-    }
-
-    private load(): void {
-        this.workspace = this.loadWorkspace();
-        this.project = this.loadProject();
-    }
-
-    private loadWorkspace(): Workspace | null {
-        const workspaceConfig = this.tree.read("/angular.json");
-        if (!workspaceConfig) return null;
-
-        return JSON.parse(workspaceConfig.toString());
-    }
-
-    private loadProject(): Project | null {
-        const projectName = this.workspace?.defaultProject;
-        if (!projectName) return null;
-
-        return this.workspace?.projects[projectName] ?? null;
+    constructor(private readonly schema: Schema, public tree: Tree) {
+        super(tree);
     }
 }
